Split index into two files
Split the index.html file into two files named indexKE.html and
indexOS.html so that we can both mess around in our own html documents
without tampering with whatever the other person might be experimenting
with. I inserted a link at the top of both pages to each other so we can
jump back and forth and see what the other person is experimenting with
without being able to alter any of the markup.

Other than that I just fixed a small error-the h1 and p tag in the
original index.html document were not in the body tag so I placed them
in there.